Changement du type d'un poids int -> double

+++ trunk/src/Algorithme/Graphe/Graphe.java
@@ -1,5 +1,7 @@
 package Algorithme.Graphe;
 
+import java.util.ArrayList;
+
 /**
  * Interface Graphe
  */
@@ -8,8 +10,8 @@ public interface Graphe {
 	public String getNom();
 	public ArrayList<Noeud> getSuivants(Noeud noeud);
 	public boolean checkTrajet(Noeud noeudDepart, Noeud noeudArrivee);
-	public int getPoids(Noeud noeudDepart, Noeud noeudArrivee);
-	public void setPoids(Noeud noeudDepart, Noeud noeudArrivee, int poids);
+	public double getPoids(Noeud noeudDepart, Noeud noeudArrivee);
+	public void setPoids(Noeud noeudDepart, Noeud noeudArrivee, double poids);
 	public int getNbreNoeuds();
 	public void setNbreNoeuds(int nbreNoeuds);
 }
